fix some warnings: field precision should have type 'int', but argument has type 'size_t'
git-svn-id: http://www.makhutov.org/svn/chan_datacard/trunk@85 2a160f14-7a49-45f3-a44f-b86709aeb494

+++ __at_parse.c
@@ -678,7 +678,7 @@ static inline int at_parse_mode (pvt_t* pvt, char* str, size_t len, int* mode, i
 
 	if (!sscanf (str, "^MODE:%d,%d", mode, submode))
 	{
-		ast_debug (2, "[%s] Error parsing MODE event '%.*s'\n", pvt->id, len, str);
+		ast_debug (2, "[%s] Error parsing MODE event '%.*s'\n", pvt->id, (int) len, str);
 		return -1;
 	}
